pyproject.toml
[build-system]
requires = ["setuptools>=66.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "aorg-directory-api"
version = "0.1"
readme = "README.md"
license = { text = "Apache-2.0" }
description = "FastAPI project"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy",
    "sqlalchemy[mypy]",
    "psycopg[binary]>=3.2.9",
    "alembic",
    "httpx>=0.27",
    "python-dotenv",
    "dishka",
    "adaptix==3.0.0b11",
    "pydantic[email]",
    "python-multipart",
    "orjson"
]
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "mypy",
    "ruff",
    "slotscheck",
    "pympler"
]
